test(table-hook): trim low-value coverage & export mock generator
