+++ .github/workflows/ci.yml
@@ -88,9 +88,6 @@ jobs:
       - name: Build
         run: pnpm build:packages
 
-      - name: Test table-hook package entry
-        run: pnpm -F @notion-kit/table-hook test:package-exports
-
       - name: Test
         run: pnpm test
