+++ packages/registry/package.json
@@ -47,6 +47,7 @@
     "@notion-kit/map": "workspace:*",
     "@notion-kit/schemas": "workspace:*",
     "@notion-kit/settings-panel": "workspace:*",
+    "@notion-kit/table-hook": "workspace:*",
     "@notion-kit/table-view": "workspace:*",
     "@notion-kit/ui": "workspace:*",
     "@notion-kit/utils": "workspace:*",
