+++ packages/table-hook/package.json
@@ -23,6 +23,10 @@
     "./plugins": {
       "types": "./dist/plugins.d.mts",
       "default": "./dist/plugins.mjs"
+    },
+    "./mock": {
+      "types": "./dist/mock.d.mts",
+      "default": "./dist/mock.mjs"
     }
   },
   "main": "./dist/index.mjs",
@@ -38,8 +42,7 @@
     "dev": "tsdown --watch",
     "format": "prettier --check . --ignore-path ../../.gitignore",
     "lint": "eslint",
-    "test": "pnpm run test:package-exports && vitest run",
-    "test:package-exports": "pnpm run build && pnpm run typecheck && node --test scripts/package-exports.test.mjs",
+    "test": "vitest run",
     "test:ui": "vitest --ui",
     "test:watch": "vitest",
     "typecheck": "tsc --noEmit --emitDeclarationOnly false"
