+++ packages/table-hook/src/__tests__/resource-api.test.tsx
@@ -1552,89 +1552,6 @@ describe("useTableView filter resource API", () => {
     expect(result.current.table.getFilters()).toBe(nextFilters);
   });
 
-  it("FilterResource_Controlled_ComposesSameRenderProposalActions", () => {
-    const onViewChange = vi.fn();
-    const proposalTwo = {
-      ...initialFilters,
-      id: "proposal-two",
-    } satisfies TableFilterState;
-    const { result } = renderHook(() =>
-      useTableView({
-        plugins,
-        defaultData: mockData,
-        defaultProperties: mockProperties,
-        view: {},
-        onViewChange,
-      }),
-    );
-
-    act(() => {
-      result.current.table.setFilters(nextFilters);
-      result.current.table.setFilters(proposalTwo);
-    });
-
-    expect(result.current.table.getFilters()).toBeUndefined();
-    expect(onViewChange).toHaveBeenCalledTimes(2);
-    expect(
-      getLastResourceChange<TableViewState, ViewResourceAction>(onViewChange)
-        ?.action,
-    ).toMatchObject({
-      type: "view.filters.change",
-      payload: {
-        previousFilters: nextFilters,
-        nextFilters: proposalTwo,
-      },
-    });
-  });
-
-  it("FilterResource_ClearFromUndefinedWritesNullThenClearingNullIsANoOp", () => {
-    const onViewChange = vi.fn();
-    const { result } = renderHook(() =>
-      useTableView({
-        plugins,
-        defaultData: mockData,
-        defaultProperties: mockProperties,
-        onViewChange,
-      }),
-    );
-
-    act(() => result.current.table.clearFilters());
-
-    expect(result.current.table.getFilters()).toBeNull();
-    expect(onViewChange).toHaveBeenCalledTimes(1);
-    expect(
-      getLastResourceChange<TableViewState, ViewResourceAction>(onViewChange)
-        ?.action,
-    ).toMatchObject({
-      type: "view.filters.change",
-      payload: { previousFilters: undefined, nextFilters: null },
-    });
-
-    act(() => result.current.table.clearFilters());
-
-    expect(result.current.table.getFilters()).toBeNull();
-    expect(onViewChange).toHaveBeenCalledTimes(1);
-  });
-
-  it("FilterResource_ExposesValidationWithoutMutatingView", () => {
-    const onViewChange = vi.fn();
-    const { result } = renderHook(() =>
-      useTableView({
-        plugins,
-        defaultData: mockData,
-        defaultProperties: mockProperties,
-        onViewChange,
-      }),
-    );
-
-    expect(result.current.table.validateFilters(nextFilters)).toBe(true);
-    expect(
-      result.current.table.validateFilters({ ...nextFilters, logic: "xor" }),
-    ).toBe(false);
-    expect(result.current.table.getFilters()).toBeUndefined();
-    expect(onViewChange).not.toHaveBeenCalled();
-  });
-
   it("FilterResource_IgnoresInvalidRuntimeSetterValues", () => {
     const onViewChange = vi.fn();
     const { result } = renderHook(() =>
