+++ packages/table-hook/src/plugins/plugins.test.ts
@@ -411,208 +411,6 @@ describe("Date filter operators", () => {
       );
     }
   });
-
-  it("does not match relative dates when either calendar conversion fails", () => {
-    const plugin = date();
-    const cases: [unknown, FilterValue, typeof config][] = [
-      [
-        { start: 1e308, includeTime: false },
-        { offsetDays: Number.MAX_SAFE_INTEGER },
-        config,
-      ],
-      [
-        { start: jan15, includeTime: false },
-        { offsetDays: 0 },
-        { ...config, tz: "Invalid/Time_Zone" },
-      ],
-    ];
-
-    for (const [data, operand, dateConfig] of cases) {
-      expect(() =>
-        matches(
-          plugin,
-          "relative-to-today",
-          data,
-          operand,
-          baseRow,
-          dateConfig,
-        ),
-      ).not.toThrow();
-      expect(
-        matches(
-          plugin,
-          "relative-to-today",
-          data,
-          operand,
-          baseRow,
-          dateConfig,
-        ),
-      ).toBe(false);
-    }
-  });
-
-  it("compares distinct pre-epoch calendar days without collapsing them", () => {
-    const plugin = date();
-    const dec30 = Date.UTC(1969, 11, 30, 12);
-    const dec31 = Date.UTC(1969, 11, 31, 12);
-
-    expect(
-      matches(plugin, "equals", { start: dec30 }, { timestamp: dec31 }),
-    ).toBe(false);
-    expect(
-      matches(plugin, "before", { start: dec30 }, { timestamp: dec31 }),
-    ).toBe(true);
-    expect(
-      matches(plugin, "after", { start: dec31 }, { timestamp: dec30 }),
-    ).toBe(true);
-    expect(
-      matches(plugin, "on-or-before", { start: dec30 }, { timestamp: dec30 }),
-    ).toBe(true);
-    expect(
-      matches(plugin, "on-or-after", { start: dec31 }, { timestamp: dec31 }),
-    ).toBe(true);
-  });
-
-  it("uses numeric instants for timed pre-epoch comparisons and ranges", () => {
-    const plugin = date();
-    const dec30 = Date.UTC(1969, 11, 30, 12);
-    const dec31 = Date.UTC(1969, 11, 31, 12);
-
-    expect(
-      matches(
-        plugin,
-        "equals",
-        { start: -1, includeTime: true },
-        { timestamp: -1 },
-      ),
-    ).toBe(true);
-    expect(
-      matches(
-        plugin,
-        "before",
-        { start: -2, includeTime: true },
-        { timestamp: -1 },
-      ),
-    ).toBe(true);
-    expect(
-      matches(plugin, "between", { start: dec31 }, { start: dec30, end: 0 }),
-    ).toBe(true);
-  });
-
-  it("compares pre-epoch calendar days in the configured timezone", () => {
-    const plugin = date();
-    const losAngeles = { ...config, tz: "America/Los_Angeles" };
-    const lateDec30Local = Date.parse("1969-12-31T07:30:00Z");
-    const noonDec30Local = Date.parse("1969-12-30T20:00:00Z");
-
-    expect(
-      matches(
-        plugin,
-        "equals",
-        { start: lateDec30Local },
-        { timestamp: noonDec30Local },
-        baseRow,
-        losAngeles,
-      ),
-    ).toBe(true);
-  });
-
-  it("uses pre-epoch row timestamps for derived date matching", () => {
-    const preEpochRow = {
-      ...baseRow,
-      createdAt: Date.UTC(1969, 11, 30, 12),
-      lastEditedAt: Date.UTC(1969, 11, 31, 12),
-    };
-
-    expect(
-      matches(
-        createdTime(),
-        "before",
-        undefined,
-        { timestamp: preEpochRow.lastEditedAt },
-        preEpochRow,
-        config,
-      ),
-    ).toBe(true);
-    expect(
-      matches(
-        lastEditedTime(),
-        "after",
-        undefined,
-        { timestamp: preEpochRow.createdAt },
-        preEpochRow,
-        config,
-      ),
-    ).toBe(true);
-  });
-
-  it("fails closed outside the negative and positive TimeClip boundaries", () => {
-    const plugin = date();
-    const outside = [-8_640_000_000_000_001, 8_640_000_000_000_001, Number.NaN];
-
-    for (const timestamp of outside) {
-      expect(() =>
-        matches(
-          plugin,
-          "equals",
-          { start: timestamp, includeTime: true },
-          { timestamp: 0 },
-        ),
-      ).not.toThrow();
-      expect(
-        matches(
-          plugin,
-          "equals",
-          { start: timestamp, includeTime: true },
-          { timestamp: 0 },
-        ),
-      ).toBe(false);
-    }
-  });
-
-  it.each([
-    ["America/Los_Angeles", false],
-    ["UTC", true],
-  ] as const)(
-    "handles exact TimeClip boundary day keys safely in %s",
-    (tz, expectedBefore) => {
-      const plugin = date();
-      const minimum = -8_640_000_000_000_000;
-      const maximum = 8_640_000_000_000_000;
-      const boundaryConfig = { ...config, tz };
-
-      expect(() =>
-        matches(
-          plugin,
-          "equals",
-          { start: minimum },
-          { timestamp: maximum },
-          baseRow,
-          boundaryConfig,
-        ),
-      ).not.toThrow();
-      expect(
-        matches(
-          plugin,
-          "equals",
-          { start: minimum },
-          { timestamp: maximum },
-          baseRow,
-          boundaryConfig,
-        ),
-      ).toBe(false);
-      expect(
-        matches(
-          plugin,
-          "before",
-          { start: minimum },
-          { timestamp: maximum },
-          baseRow,
-          boundaryConfig,
-        ),
-      ).toBe(expectedBefore);
-    },
-  );
 });
 
 describe("configured plugin factories", () => {
